+++ src/Monolog/Handler.php
@@ -69,8 +69,14 @@ final class Handler extends AbstractProcessingHandler
 
             $monologContextData = $this->getMonologContextData($record['context']);
 
-            if (!empty($monologContextData)) {
+            if ([] !== $monologContextData) {
                 $scope->setExtra('monolog.context', $monologContextData);
+            }
+
+            $monologExtraData = $this->getMonologExtraData($record['extra']);
+
+            if ([] !== $monologExtraData) {
+                $scope->setExtra('monolog.extra', $monologExtraData);
             }
 
             $this->hub->captureEvent($event, $hint);
@@ -100,5 +106,25 @@ final class Handler extends AbstractProcessingHandler
         }
 
         return $contextData;
+    }
+
+    /**
+     * @param mixed[] $context
+     *
+     * @return mixed[]
+     */
+    private function getMonologExtraData(array $context): array
+    {
+        if (!$this->fillExtraContext) {
+            return [];
+        }
+
+        $extraData = [];
+
+        foreach ($context as $key => $value) {
+            $extraData[$key] = $value;
+        }
+
+        return $extraData;
     }
 }